Refactor run description and status tests following review.
