+++ backend/entityservice/tests/conftest.py
@@ -1,15 +1,23 @@
 import time
 import pytest
 import requests as requests_library
+
 THROTTLE_SLEEP = 0.1
 
 
 @pytest.fixture
 def requests():
     """
-    We inject the requests library so that we could potentially monkey patch it
-    For now we just add a small sleep to ensure we don't get throttled when
-    tests run back to back.
+    We inject the requests session.
+    For now we just add a small sleep after every request to ensure we don't get throttled when
+    tests run back to back. Note the rate limit in nginx is 10 requests per ip per second.
     """
-    yield requests_library
-    time.sleep(THROTTLE_SLEEP)
+
+    testing_session = requests_library.Session()
+
+    def delay_next(r, *args, **kwargs):
+        time.sleep(THROTTLE_SLEEP)
+
+    testing_session.hooks['response'].append(delay_next)
+
+    yield testing_session
